todo state hook 으로 분리
